requirements.txt
Django>=4.0,<4.1
Pillow>=8.4.0,<8.5
pytesseract